Avoid modulizing filename when reading file
The original behavior was to modulize the filename in Read.file
as a convenience to the caller.
Then the caller Unit.read_file would ignore the modulize filename.

This behavior is a bug because the codept user may have supplied
a valid Namespaced.t for the file. The result would be:

     Invalid_argument("Impossible to modulize \"_init\" (from \"_init.ml\")")
     Raised at Stdlib.invalid_arg in file "stdlib.ml", line 30, characters 20-45
     Called from Unitname.modulize in file "lib/unitname.ml", line 16, characters 2-109
     Called from Read.file in file "lib/read.ml", line 49, characters 13-26
     Called from Unit.read_file in file "lib/unit.ml", line 49, characters 20-43

1. Now a new Read.file_raw function is available that avoids the
modulizing of the filename.

2. These callers do not need the modulizing so they have switched to Read.file_raw:
- Unit.read_file
- Single.to_m2l